Package "webhook" was renamed to "bot".Package structure was changed, code refactor.
