+++ src/main/java/com/softserve/bot/WebhookBotStarter.java
@@ -1,0 +1,11 @@
+package com.softserve.bot;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class WebhookBotStarter {
+    public static void main(String[] args) {
+        SpringApplication.run(WebhookBotStarter.class, args);
+    }
+}
