+++ src/main/java/com/softserve/bot/view/handler/SpecializationHandler.java
@@ -1,0 +1,55 @@
+package com.softserve.bot.view.handler;
+
+import com.softserve.bot.model.Subject;
+import com.softserve.bot.service.Filter;
+import com.softserve.bot.model.Specialty;
+import lombok.AllArgsConstructor;
+import org.springframework.stereotype.Component;
+import org.telegram.telegrambots.meta.api.methods.send.SendMessage;
+import org.telegram.telegrambots.meta.api.objects.Message;
+import org.telegram.telegrambots.meta.api.objects.Update;
+
+import java.util.EnumSet;
+import java.util.List;
+import java.util.Map;
+
+@AllArgsConstructor
+@Component
+public class SpecializationHandler {
+    private Filter filter;
+    private Map<String, List<Specialty>> filteredSpecialty;
+    private Message message;
+    private SendMessage sendMessage;
+
+    private void cleanRequests() {
+        sendMessage.setReplyMarkup(null);
+    }
+
+    public SendMessage handle(Update update) {
+        cleanRequests();
+        if (update.getCallbackQuery() != null) {
+            sendMessage.setChatId(String.valueOf(update.getCallbackQuery().getMessage().getChatId()));
+            // прикрепть кнопки выбора специальности
+        } else {
+            sendMessage.setChatId(String.valueOf(update.getMessage().getChatId()));
+        }
+        sendMessage.setText("Cпеціальності:1");
+        return sendMessage;
+    }
+
+    public SendMessage handleFiltered(Update update, EnumSet<Subject> enumSet) {
+        message = update.getMessage();
+        sendMessage.setChatId(String.valueOf(message.getChatId()));
+        filteredSpecialty = filter.getFiltered(enumSet);
+        StringBuilder stringBuilder = new StringBuilder();
+        for (List<Specialty> value : filteredSpecialty.values()) {
+            for (Specialty currentSpecialty : value) {
+                stringBuilder.append(currentSpecialty).append("\n");
+            }
+        }
+        // прикрепть кнопки выбора специальности
+        sendMessage.setText(String.valueOf(stringBuilder));
+        return sendMessage;
+    }
+
+}
