+++ src/main/java/com/softserve/bot/view/handler/SubjectHandler.java
@@ -1,0 +1,73 @@
+package com.softserve.bot.view.handler;
+
+import com.softserve.bot.util.ButtonSubjectRegister;
+import com.softserve.bot.view.sender.AlertSender;
+import com.softserve.bot.model.Subject;
+import com.softserve.bot.util.RadioButton;
+import lombok.AllArgsConstructor;
+import org.springframework.stereotype.Component;
+import org.telegram.telegrambots.meta.api.methods.BotApiMethod;
+import org.telegram.telegrambots.meta.api.methods.send.SendMessage;
+import org.telegram.telegrambots.meta.api.methods.updatingmessages.EditMessageReplyMarkup;
+import org.telegram.telegrambots.meta.api.objects.Message;
+import org.telegram.telegrambots.meta.api.objects.Update;
+
+import java.io.Serializable;
+import java.util.EnumSet;
+
+@AllArgsConstructor
+@Component
+public class SubjectHandler {
+    private ButtonSubjectRegister buttonSubjectRegister;
+    private EditMessageReplyMarkup editMessageReplyMarkup;
+    private Message message;
+    private SendMessage sendMessage;
+    private AlertSender alertSender;
+
+    private void cleanRequests() {
+        sendMessage.setReplyMarkup(null);
+    }
+
+    public EditMessageReplyMarkup deleteSelectedSubject(Update update, EnumSet<Subject> enumSet) {
+        enumSet = EnumSet.of(Subject.UKRAINIAN, Subject.MATH_PROFILE);
+        return processingSelectionSubject(update, enumSet);
+    }
+
+    private EditMessageReplyMarkup processingSelectionSubject(Update update, EnumSet<Subject> enumSet) {
+        editMessageReplyMarkup.setChatId(String.valueOf(update.getCallbackQuery().getMessage().getChatId()));
+        editMessageReplyMarkup.setMessageId(update.getCallbackQuery().getMessage().getMessageId());
+        editMessageReplyMarkup.setReplyMarkup(buttonSubjectRegister.getInlineSubjectButtons(enumSet, enumSet.size()));
+        return editMessageReplyMarkup;
+    }
+
+    public SendMessage handle(Update update, EnumSet<Subject> enumSet) {
+        cleanRequests();
+        message = update.getMessage();
+        sendMessage.setChatId(String.valueOf(message.getChatId()));
+        sendMessage.setReplyMarkup(buttonSubjectRegister.getInlineSubjectButtons(enumSet, enumSet.size()));
+        sendMessage.setText("Усі предмети:");
+        return sendMessage;
+    }
+
+    public BotApiMethod<? extends Serializable> setAndRemoveTick(Update update, Subject element, EnumSet<Subject> enumSet) {
+        if (enumSet.contains(element)) {
+            if (RadioButton.notOutOfLimit(enumSet)) {
+                RadioButton.removeMandatoryTick(element, enumSet);
+                return processingSelectionSubject(update, enumSet);
+            } else {
+                return alertSender.sendSubjectAlert(update);
+            }
+        } else if (RadioButton.notOutOfLimit(enumSet)) {
+            RadioButton.radioButtonImpl(element, enumSet);
+            return processingSelectionSubject(update, enumSet);
+        } else {
+            if (RadioButton.outOfLimitChecker(element, enumSet)) {
+                RadioButton.radioButtonImpl(element, enumSet);
+                return processingSelectionSubject(update, enumSet);
+            }
+            return alertSender.sendSubjectAlert(update);
+        }
+    }
+
+
+}
