--- src/main/java/com/softserve/webhookbot/Specialty.java
@@ -1,30 +1,0 @@
-package com.softserve.webhookbot;
-
-import com.softserve.webhookbot.enumeration.Subject;
-
-import java.util.EnumSet;
-
-@Deprecated(forRemoval = true)
-public class Specialty {
-    private String code;
-    private String name;
-
-    private EnumSet<Subject> subjects;
-
-    public String getName() {
-        return name;
-    }
-
-    public String getCode() {
-        return code;
-    }
-
-    public void setSpecialties(EnumSet<Subject> subjects) {
-        this.subjects = subjects;
-    }
-
-    public EnumSet<Subject> getSubjects() {
-        return subjects;
-    }
-
-}
