--- src/main/java/com/softserve/webhookbot/util/UpdateSessionParser.java
@@ -1,28 +1,0 @@
-package com.softserve.webhookbot.util;
-
-import com.softserve.webhookbot.enumeration.Subject;
-import org.springframework.stereotype.Component;
-import org.telegram.telegrambots.meta.api.objects.Update;
-
-import java.util.EnumSet;
-
-@Component
-public class UpdateSessionParser {
-    private EnumSet<Subject> enumSet;
-    private String callback;
-
-
-    public void parse(Update update) {
-        String[] data = update.getCallbackQuery().getData().split("/");
-        enumSet = EnumSetUtil.decode(Integer.parseInt(data[1]), Subject.class);
-        callback =data[0];
-    }
-
-    public EnumSet<Subject> getEnumSet() {
-        return enumSet;
-    }
-
-    public String getCallback() {
-        return callback;
-    }
-}
