+++ src/test/java/com/softserve/bot/WebhookBotStarterTests.java
@@ -1,0 +1,13 @@
+package com.softserve.bot;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class WebhookBotStarterTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
